+++ py-bin/tlbin/__init__.py
@@ -1,0 +1,1 @@
+from .contracts import load_packaged_contracts  # noqa: F401